reference-app/frontend/requirements.txt
Flask
Flask-OpenTracing
requests
pylint
gunicorn
prometheus-client
opentelemetry-instrumentation
opentelemetry-instrumentation-requests
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-flask
prometheus-flask-exporter
jaeger-client
opentelemetry-api

